Dispose USB resources
Sometimes the USB spectrometers can't be reused after the script
finished or crashed. This commit adds a dispose() method to all
Spectrometer classes, that can either be called in the finally
block of a try-except statement or the whole class can be used
in a with statement.

Thanks go to Axel Fischer.